# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Brain Bridge API Documentation",
        "description": "API documentation for Brain Bridge application",
        "contact": {
            "email": "gmzesan7767@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/teacher-levels": {
            "get": {
                "tags": [
                    "Teacher Levels"
                ],
                "summary": "Get all teacher levels",
                "description": "Returns a list of all teacher levels",
                "operationId": "getTeacherLevels",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Teacher Levels"
                ],
                "summary": "Create a new teacher level",
                "operationId": "storeTeacherLevel",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "level_name"
                                ],
                                "properties": {
                                    "level_name": {
                                        "type": "string",
                                        "example": "Senior"
                                    },
                                    "min_rating": {
                                        "type": "integer",
                                        "example": 50
                                    },
                                    "max_rating": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "benefits": {
                                        "type": "string",
                                        "example": "Access to premium resources"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Teacher level created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/teacher-levels/{id}": {
            "get": {
                "tags": [
                    "Teacher Levels"
                ],
                "summary": "Get a specific teacher level",
                "operationId": "showTeacherLevel",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Teacher Level ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Teacher Level not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Teacher Levels"
                ],
                "summary": "Update a teacher level",
                "operationId": "updateTeacherLevel",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Teacher Level ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "level_name": {
                                        "type": "string",
                                        "example": "Senior"
                                    },
                                    "min_rating": {
                                        "type": "integer",
                                        "example": 50
                                    },
                                    "max_rating": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "benefits": {
                                        "type": "string",
                                        "example": "Access to premium resources"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Teacher level updated successfully"
                    },
                    "404": {
                        "description": "Teacher Level not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Teacher Levels"
                ],
                "summary": "Delete a teacher level",
                "operationId": "deleteTeacherLevel",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Teacher Level ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Teacher level deleted successfully"
                    },
                    "404": {
                        "description": "Teacher Level not found"
                    }
                }
            }
        },
...
    },
...
}